requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-rest-swagger==2.2.0
psycopg2==2.9.3
PyYAML==6.0
drf-yasg==1.20.0